hardware-services/requirements.txt
requests
pymongo
python-dotenv